+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (command_permission module) feature: add `/command-permission verbose` command, to enable verbose-mode, and see what is happening in console.
-- (command_permission module) refactor: rename `/command-permission` to `/command-permission gui` command.
+- (command_debug module) feature: add `/command-debug <command>` command, to see debug info of a command.